test: Unittests for 2.0.0 (#28)
* test: get_tokenizer tests

* fix: cases

* test: Added some unittests to resources

* fix: rename var

* test: Added ai21 studio client tsts

* fix: rename files

* fix: Added types

* test: added test to http

* fix: removed unnecessary auth param

* test: Added tests

* test: Added sagemaker

* test: Created a single session per instance

* ci: removed unnecessary action

+++ ai21/clients/studio/resources/studio_completion.py
@@ -15,7 +15,7 @@ class StudioCompletion(StudioResource, Completion):
         num_results: Optional[int] = 1,
         min_tokens: Optional[int] = 0,
         temperature: Optional[float] = 0.7,
-        top_p: Optional[int] = 1,
+        top_p: Optional[float] = 1,
         top_k_return: Optional[int] = 0,
         custom_model: Optional[str] = None,
         experimental_mode: bool = False,
